컴포넌트 기능별로 파일 분리
